Moves safe browsing key to native for Android
